model: event.rs: Implement a getter for `prev_events`
This implements a method which allows us to get the `event_id` of
events preceding another. This is necessary because on some versions
of rooms, `prev_events` can be the events' IDs and their hashes
or their IDs alone. As it is not possible to derive serialization
traits from serde which would allow two different types for this field,
we just type it as a generic `serde_json::Value` and then extract the
information we need thanks to this method.