dag.rs: Properly manipulate the events DAG
This makes corrections so that the DAG is created and events are
correctly added to it.
Now it takes into account that there isn't just one earlier/latest event
to the DAG but possibly several of them.
We can also now create a correct data set (for vis.js) from the DAG and
retrieve events which are older/newer than a set of events. Which is
useful for properly updating the displayed DAG without having to
completely redraw it.